+++ juniper/src/executor/mod.rs
@@ -24,8 +24,10 @@ use crate::{
         model::{RootNode, SchemaType, TypeType},
     },
     types::{
-        async_await::GraphQLTypeAsync, base::GraphQLType, name::Name,
-        subscriptions::GraphQLSubscriptionType,
+        async_await::{GraphQLTypeAsync, GraphQLValueAsync},
+        base::{GraphQLType, GraphQLValue},
+        name::Name,
+        subscriptions::{GraphQLSubscriptionType, GraphQLSubscriptionValue},
     },
     value::{DefaultScalarValue, ParseScalarValue, ScalarValue, Value},
     GraphQLError,
@@ -248,7 +250,7 @@ impl<S> IntoFieldError<S> for FieldError<S> {
 #[doc(hidden)]
 pub trait IntoResolvable<'a, S, T, C>
 where
-    T: GraphQLType<S>,
+    T: GraphQLValue<S>,
     S: ScalarValue,
 {
     #[doc(hidden)]
@@ -257,7 +259,7 @@ where
 
 impl<'a, S, T, C> IntoResolvable<'a, S, T, C> for T
 where
-    T: GraphQLType<S>,
+    T: GraphQLValue<S>,
     S: ScalarValue,
     T::Context: FromContext<C>,
 {
@@ -269,7 +271,7 @@ where
 impl<'a, S, T, C, E: IntoFieldError<S>> IntoResolvable<'a, S, T, C> for Result<T, E>
 where
     S: ScalarValue,
-    T: GraphQLType<S>,
+    T: GraphQLValue<S>,
     T::Context: FromContext<C>,
 {
     fn into(self, ctx: &'a C) -> FieldResult<Option<(&'a T::Context, T)>, S> {
@@ -281,7 +283,7 @@ where
 impl<'a, S, T, C> IntoResolvable<'a, S, T, C> for (&'a T::Context, T)
 where
     S: ScalarValue,
-    T: GraphQLType<S>,
+    T: GraphQLValue<S>,
 {
     fn into(self, _: &'a C) -> FieldResult<Option<(&'a T::Context, T)>, S> {
         Ok(Some(self))
@@ -291,7 +293,7 @@ where
 impl<'a, S, T, C> IntoResolvable<'a, S, Option<T>, C> for Option<(&'a T::Context, T)>
 where
     S: ScalarValue,
-    T: GraphQLType<S>,
+    T: GraphQLValue<S>,
 {
     fn into(self, _: &'a C) -> FieldResult<Option<(&'a T::Context, Option<T>)>, S> {
         Ok(self.map(|(ctx, v)| (ctx, Some(v))))
@@ -301,7 +303,7 @@ where
 impl<'a, S, T, C> IntoResolvable<'a, S, T, C> for FieldResult<(&'a T::Context, T), S>
 where
     S: ScalarValue,
-    T: GraphQLType<S>,
+    T: GraphQLValue<S>,
 {
     fn into(self, _: &'a C) -> FieldResult<Option<(&'a T::Context, T)>, S> {
         self.map(Some)
@@ -312,7 +314,7 @@ impl<'a, S, T, C> IntoResolvable<'a, S, Option<T>, C>
     for FieldResult<Option<(&'a T::Context, T)>, S>
 where
     S: ScalarValue,
-    T: GraphQLType<S>,
+    T: GraphQLValue<S>,
 {
     fn into(self, _: &'a C) -> FieldResult<Option<(&'a T::Context, Option<T>)>, S> {
         self.map(|o| o.map(|(ctx, v)| (ctx, Some(v))))
@@ -373,7 +375,7 @@ where
         'i: 'res,
         'v: 'res,
         'a: 'res,
-        T: GraphQLSubscriptionType<S, Context = CtxT> + Sync,
+        T: GraphQLSubscriptionValue<S, Context = CtxT>,
         T::TypeInfo: Sync,
         CtxT: Sync,
         S: Send + Sync,
@@ -397,7 +399,7 @@ where
     where
         't: 'res,
         'a: 'res,
-        T: GraphQLSubscriptionType<S, Context = CtxT>,
+        T: GraphQLSubscriptionValue<S, Context = CtxT> + ?Sized,
         T::TypeInfo: Sync,
         CtxT: Sync,
         S: Send + Sync,
@@ -409,7 +411,7 @@ where
     pub fn resolve_with_ctx<NewCtxT, T>(&self, info: &T::TypeInfo, value: &T) -> ExecutionResult<S>
     where
         NewCtxT: FromContext<CtxT>,
-        T: GraphQLType<S, Context = NewCtxT> + ?Sized,
+        T: GraphQLValue<S, Context = NewCtxT> + ?Sized,
     {
         self.replaced_context(<NewCtxT as FromContext<CtxT>>::from(self.context))
             .resolve(info, value)
@@ -418,7 +420,7 @@ where
     /// Resolve a single arbitrary value into an `ExecutionResult`
     pub fn resolve<T>(&self, info: &T::TypeInfo, value: &T) -> ExecutionResult<S>
     where
-        T: GraphQLType<S, Context = CtxT> + ?Sized,
+        T: GraphQLValue<S, Context = CtxT> + ?Sized,
     {
         value.resolve(info, self.current_selection_set, self)
     }
@@ -426,7 +428,7 @@ where
     /// Resolve a single arbitrary value into an `ExecutionResult`
     pub async fn resolve_async<T>(&self, info: &T::TypeInfo, value: &T) -> ExecutionResult<S>
     where
-        T: GraphQLTypeAsync<S, Context = CtxT> + ?Sized,
+        T: GraphQLValueAsync<S, Context = CtxT> + ?Sized,
         T::TypeInfo: Sync,
         CtxT: Sync,
         S: Send + Sync,
@@ -443,7 +445,7 @@ where
         value: &T,
     ) -> ExecutionResult<S>
     where
-        T: GraphQLTypeAsync<S, Context = NewCtxT>,
+        T: GraphQLValueAsync<S, Context = NewCtxT>,
         T::TypeInfo: Sync,
         NewCtxT: FromContext<CtxT> + Sync,
         S: Send + Sync,
@@ -457,15 +459,12 @@ where
     /// If the field fails to resolve, `null` will be returned.
     pub fn resolve_into_value<T>(&self, info: &T::TypeInfo, value: &T) -> Value<S>
     where
-        T: GraphQLType<S, Context = CtxT>,
+        T: GraphQLValue<S, Context = CtxT>,
     {
-        match self.resolve(info, value) {
-            Ok(v) => v,
-            Err(e) => {
-                self.push_error(e);
-                Value::null()
-            }
-        }
+        self.resolve(info, value).unwrap_or_else(|e| {
+            self.push_error(e);
+            Value::null()
+        })
     }
 
     /// Resolve a single arbitrary value into a return value
@@ -473,7 +472,7 @@ where
     /// If the field fails to resolve, `null` will be returned.
     pub async fn resolve_into_value_async<T>(&self, info: &T::TypeInfo, value: &T) -> Value<S>
     where
-        T: GraphQLTypeAsync<S, Context = CtxT> + ?Sized,
+        T: GraphQLValueAsync<S, Context = CtxT> + ?Sized,
         T::TypeInfo: Sync,
         CtxT: Sync,
         S: Send + Sync,