+++ juniper/src/lib.rs
@@ -93,6 +93,10 @@ Juniper has not reached 1.0 yet, thus some API instability should be expected.
 #![doc(html_root_url = "https://docs.rs/juniper/0.14.2")]
 #![warn(missing_docs)]
 
+// Required for using `juniper_codegen` macros inside this crate to resolve absolute `::juniper`
+// path correctly, without errors.
+extern crate self as juniper;
+
 use std::fmt;
 
 #[doc(hidden)]
@@ -126,14 +130,6 @@ pub use futures::future::BoxFuture;
 pub use juniper_codegen::{
     graphql_object, graphql_scalar, graphql_subscription, graphql_union, GraphQLEnum,
     GraphQLInputObject, GraphQLObject, GraphQLScalarValue, GraphQLUnion,
-};
-// Internal macros are not exported,
-// but declared at the root to make them easier to use.
-#[allow(unused_imports)]
-use juniper_codegen::{
-    graphql_object_internal, graphql_scalar_internal, graphql_subscription_internal,
-    graphql_union_internal, GraphQLEnumInternal, GraphQLInputObjectInternal,
-    GraphQLScalarValueInternal, GraphQLUnionInternal,
 };
 
 #[macro_use]
@@ -185,11 +181,14 @@ pub use crate::{
         model::{RootNode, SchemaType},
     },
     types::{
-        async_await::GraphQLTypeAsync,
-        base::{Arguments, GraphQLType, TypeKind},
+        async_await::{GraphQLTypeAsync, GraphQLValueAsync},
+        base::{Arguments, GraphQLType, GraphQLValue, TypeKind},
         marker::{self, GraphQLUnion},
         scalars::{EmptyMutation, EmptySubscription, ID},
-        subscriptions::{GraphQLSubscriptionType, SubscriptionConnection, SubscriptionCoordinator},
+        subscriptions::{
+            GraphQLSubscriptionType, GraphQLSubscriptionValue, SubscriptionConnection,
+            SubscriptionCoordinator,
+        },
     },
     validation::RuleError,
     value::{DefaultScalarValue, Object, ParseScalarResult, ParseScalarValue, ScalarValue, Value},