+++ juniper/src/types/pointers.rs
@@ -1,12 +1,12 @@
-use std::{fmt::Debug, sync::Arc};
+use std::{fmt, sync::Arc};
 
 use crate::{
     ast::{FromInputValue, InputValue, Selection, ToInputValue},
     executor::{ExecutionResult, Executor, Registry},
     schema::meta::MetaType,
     types::{
-        async_await::GraphQLTypeAsync,
-        base::{Arguments, GraphQLType},
+        async_await::GraphQLValueAsync,
+        base::{Arguments, GraphQLType, GraphQLValue},
     },
     value::ScalarValue,
     BoxFuture,
@@ -17,9 +17,6 @@ where
     T: GraphQLType<S, Context = CtxT> + ?Sized,
     S: ScalarValue,
 {
-    type Context = CtxT;
-    type TypeInfo = T::TypeInfo;
-
     fn name(info: &T::TypeInfo) -> Option<&str> {
         T::name(info)
     }
@@ -29,6 +26,19 @@ where
         S: 'r,
     {
         T::meta(info, registry)
+    }
+}
+
+impl<S, T, CtxT> GraphQLValue<S> for Box<T>
+where
+    S: ScalarValue,
+    T: GraphQLValue<S, Context = CtxT> + ?Sized,
+{
+    type Context = CtxT;
+    type TypeInfo = T::TypeInfo;
+
+    fn type_name<'i>(&self, info: &'i T::TypeInfo) -> Option<&'i str> {
+        (**self).type_name(info)
     }
 
     fn resolve_into_type(
@@ -61,11 +71,11 @@ where
     }
 }
 
-impl<S, T, CtxT> GraphQLTypeAsync<S> for Box<T>
+impl<S, T> GraphQLValueAsync<S> for Box<T>
 where
-    T: GraphQLTypeAsync<S, Context = CtxT> + ?Sized,
+    T: GraphQLValueAsync<S> + ?Sized,
     T::TypeInfo: Sync,
-    CtxT: Sync,
+    T::Context: Sync,
     S: ScalarValue + Send + Sync,
 {
     fn resolve_async<'a>(
@@ -93,7 +103,7 @@ where
 
 impl<T, S> ToInputValue<S> for Box<T>
 where
-    S: Debug,
+    S: fmt::Debug,
     T: ToInputValue<S>,
 {
     fn to_input_value(&self) -> InputValue<S> {
@@ -106,9 +116,6 @@ where
     T: GraphQLType<S, Context = CtxT> + ?Sized,
     S: ScalarValue,
 {
-    type Context = CtxT;
-    type TypeInfo = T::TypeInfo;
-
     fn name(info: &T::TypeInfo) -> Option<&str> {
         T::name(info)
     }
@@ -118,6 +125,19 @@ where
         S: 'r,
     {
         T::meta(info, registry)
+    }
+}
+
+impl<'e, S, T, CtxT> GraphQLValue<S> for &'e T
+where
+    S: ScalarValue,
+    T: GraphQLValue<S, Context = CtxT> + ?Sized,
+{
+    type Context = CtxT;
+    type TypeInfo = T::TypeInfo;
+
+    fn type_name<'i>(&self, info: &'i T::TypeInfo) -> Option<&'i str> {
+        (**self).type_name(info)
     }
 
     fn resolve_into_type(
@@ -150,9 +170,9 @@ where
     }
 }
 
-impl<'e, S, T> GraphQLTypeAsync<S> for &'e T
+impl<'e, S, T> GraphQLValueAsync<S> for &'e T
 where
-    T: GraphQLTypeAsync<S> + ?Sized,
+    T: GraphQLValueAsync<S> + ?Sized,
     T::TypeInfo: Sync,
     T::Context: Sync,
     S: ScalarValue + Send + Sync,
@@ -164,7 +184,7 @@ where
         arguments: &'b Arguments<S>,
         executor: &'b Executor<Self::Context, S>,
     ) -> BoxFuture<'b, ExecutionResult<S>> {
-        GraphQLTypeAsync::resolve_field_async(&**self, info, field_name, arguments, executor)
+        (**self).resolve_field_async(info, field_name, arguments, executor)
     }
 
     fn resolve_async<'a>(
@@ -173,13 +193,13 @@ where
         selection_set: Option<&'a [Selection<S>]>,
         executor: &'a Executor<Self::Context, S>,
     ) -> BoxFuture<'a, ExecutionResult<S>> {
-        GraphQLTypeAsync::resolve_async(&**self, info, selection_set, executor)
+        (**self).resolve_async(info, selection_set, executor)
     }
 }
 
 impl<'a, T, S> ToInputValue<S> for &'a T
 where
-    S: Debug,
+    S: fmt::Debug,
     T: ToInputValue<S>,
 {
     fn to_input_value(&self) -> InputValue<S> {
@@ -192,9 +212,6 @@ where
     S: ScalarValue,
     T: GraphQLType<S> + ?Sized,
 {
-    type Context = T::Context;
-    type TypeInfo = T::TypeInfo;
-
     fn name(info: &T::TypeInfo) -> Option<&str> {
         T::name(info)
     }
@@ -204,6 +221,19 @@ where
         S: 'r,
     {
         T::meta(info, registry)
+    }
+}
+
+impl<S, T> GraphQLValue<S> for Arc<T>
+where
+    S: ScalarValue,
+    T: GraphQLValue<S> + ?Sized,
+{
+    type Context = T::Context;
+    type TypeInfo = T::TypeInfo;
+
+    fn type_name<'i>(&self, info: &'i T::TypeInfo) -> Option<&'i str> {
+        (**self).type_name(info)
     }
 
     fn resolve_into_type(
@@ -236,9 +266,9 @@ where
     }
 }
 
-impl<'e, S, T> GraphQLTypeAsync<S> for Arc<T>
+impl<'e, S, T> GraphQLValueAsync<S> for Arc<T>
 where
-    T: GraphQLTypeAsync<S> + Send + ?Sized,
+    T: GraphQLValueAsync<S> + Send + ?Sized,
     T::TypeInfo: Sync,
     T::Context: Sync,
     S: ScalarValue + Send + Sync,
@@ -255,7 +285,7 @@ where
 
 impl<T, S> ToInputValue<S> for Arc<T>
 where
-    S: Debug,
+    S: fmt::Debug,
     T: ToInputValue<S>,
 {
     fn to_input_value(&self) -> InputValue<S> {