+++ juniper/src/types/subscriptions.rs
@@ -3,9 +3,9 @@ use futures::{future, stream};
 use crate::{
     http::{GraphQLRequest, GraphQLResponse},
     parser::Spanning,
-    types::base::{is_excluded, merge_key_into},
-    Arguments, BoxFuture, Executor, FieldError, GraphQLType, Object, ScalarValue, Selection, Value,
-    ValuesStream,
+    types::base::{is_excluded, merge_key_into, GraphQLType, GraphQLValue},
+    Arguments, BoxFuture, DefaultScalarValue, Executor, FieldError, Object, ScalarValue, Selection,
+    Value, ValuesStream,
 };
 
 /// Global subscription coordinator trait.
@@ -60,23 +60,24 @@ where
 /// server integration crates.
 pub trait SubscriptionConnection<'a, S>: futures::Stream<Item = GraphQLResponse<'a, S>> {}
 
-/**
- This trait adds resolver logic with asynchronous subscription execution logic
- on GraphQL types. It should be used with `GraphQLType` in order to implement
- subscription resolvers on GraphQL objects.
-
- Subscription-related convenience macros expand into an implementation of this
- trait and `GraphQLType` for the given type.
-
- See trait methods for more detailed explanation on how this trait works.
-*/
-pub trait GraphQLSubscriptionType<S>: GraphQLType<S> + Sync
+/// Extension of [`GraphQLValue`] trait with asynchronous [subscription][1] execution logic.
+/// It should be used with [`GraphQLValue`] in order to implement [subscription][1] resolvers on
+/// [GraphQL objects][2].
+///
+/// [Subscription][1]-related convenience macros expand into an implementation of this trait and
+/// [`GraphQLValue`] for the given type.
+///
+/// See trait methods for more detailed explanation on how this trait works.
+///
+/// [1]: https://spec.graphql.org/June2018/#sec-Subscription
+/// [2]: https://spec.graphql.org/June2018/#sec-Objects
+pub trait GraphQLSubscriptionValue<S = DefaultScalarValue>: GraphQLValue<S> + Sync
 where
     Self::TypeInfo: Sync,
     Self::Context: Sync,
     S: ScalarValue + Send + Sync,
 {
-    /// Resolve into `Value<ValuesStream>`
+    /// Resolves into `Value<ValuesStream>`.
     ///
     /// ## Default implementation
     ///
@@ -111,7 +112,7 @@ where
     /// This method is called by Self's `resolve_into_stream` default
     /// implementation every time any field is found in selection set.
     ///
-    /// It replaces `GraphQLType::resolve_field`.
+    /// It replaces `GraphQLValue::resolve_field`.
     /// Unlike `resolve_field`, which resolves each field into a single
     /// `Value<S>`, this method resolves each field into
     /// `Value<ValuesStream<S>>`.
@@ -140,7 +141,7 @@ where
     /// This method is called by Self's `resolve_into_stream` default
     /// implementation every time any fragment is found in selection set.
     ///
-    /// It replaces `GraphQLType::resolve_into_type`.
+    /// It replaces `GraphQLValue::resolve_into_type`.
     /// Unlike `resolve_into_type`, which resolves each fragment
     /// a single `Value<S>`, this method resolves each fragment into
     /// `Value<ValuesStream<S>>`.
@@ -162,13 +163,39 @@ where
         'res: 'f,
     {
         Box::pin(async move {
-            if Self::name(info) == Some(type_name) {
+            if self.type_name(info) == Some(type_name) {
                 self.resolve_into_stream(info, executor).await
             } else {
                 panic!("resolve_into_type_stream must be implemented");
             }
         })
     }
+}
+
+crate::sa::assert_obj_safe!(GraphQLSubscriptionValue<Context = (), TypeInfo = ()>);
+
+/// Extension of [`GraphQLType`] trait with asynchronous [subscription][1] execution logic.
+///
+/// It's automatically implemented for [`GraphQLSubscriptionValue`] and [`GraphQLType`]
+/// implementors, so doesn't require manual or code-generated implementation.
+///
+/// [1]: https://spec.graphql.org/June2018/#sec-Subscription
+pub trait GraphQLSubscriptionType<S = DefaultScalarValue>:
+    GraphQLSubscriptionValue<S> + GraphQLType<S>
+where
+    Self::Context: Sync,
+    Self::TypeInfo: Sync,
+    S: ScalarValue + Send + Sync,
+{
+}
+
+impl<S, T> GraphQLSubscriptionType<S> for T
+where
+    T: GraphQLSubscriptionValue<S> + GraphQLType<S>,
+    T::Context: Sync,
+    T::TypeInfo: Sync,
+    S: ScalarValue + Send + Sync,
+{
 }
 
 /// Wrapper function around `resolve_selection_set_into_stream_recursive`.
@@ -186,7 +213,7 @@ where
     'e: 'fut,
     'ref_e: 'fut,
     'res: 'fut,
-    T: GraphQLSubscriptionType<S> + ?Sized,
+    T: GraphQLSubscriptionValue<S> + ?Sized,
     T::TypeInfo: Sync,
     T::Context: Sync,
     S: ScalarValue + Send + Sync,
@@ -205,7 +232,7 @@ async fn resolve_selection_set_into_stream_recursive<'i, 'inf, 'ref_e, 'e, 'res,
     executor: &'ref_e Executor<'ref_e, 'e, T::Context, S>,
 ) -> Value<ValuesStream<'res, S>>
 where
-    T: GraphQLSubscriptionType<S> + ?Sized,
+    T: GraphQLSubscriptionValue<S> + ?Sized,
     T::TypeInfo: Sync,
     T::Context: Sync,
     S: ScalarValue + Send + Sync,
@@ -220,7 +247,8 @@ where
     let meta_type = executor
         .schema()
         .concrete_type_by_name(
-            T::name(info)
+            instance
+                .type_name(info)
                 .expect("Resolving named type's selection set")
                 .as_ref(),
         )