+++ juniper/src/value/scalar.rs
@@ -1,7 +1,11 @@
-use crate::parser::{ParseError, ScalarToken};
-use juniper_codegen::GraphQLScalarValueInternal as GraphQLScalarValue;
+use std::fmt;
+
 use serde::{de, ser::Serialize};
-use std::fmt::{self, Debug, Display};
+
+use crate::{
+    parser::{ParseError, ScalarToken},
+    GraphQLScalarValue,
+};
 
 /// The result of converting a string into a scalar value
 pub type ParseScalarResult<'a, S = DefaultScalarValue> = Result<S, ParseError<'a>>;
@@ -164,7 +168,15 @@ pub trait ParseScalarValue<S = DefaultScalarValue> {
 /// # fn main() {}
 /// ```
 pub trait ScalarValue:
-    Debug + Display + PartialEq + Clone + Serialize + From<String> + From<bool> + From<i32> + From<f64>
+    fmt::Debug
+    + fmt::Display
+    + PartialEq
+    + Clone
+    + Serialize
+    + From<String>
+    + From<bool>
+    + From<i32>
+    + From<f64>
 {
     /// Serde visitor used to deserialize this scalar value
     type Visitor: for<'de> de::Visitor<'de, Value = Self> + Default;
@@ -190,26 +202,26 @@ pub trait ScalarValue:
 
     /// Convert the given scalar value into an integer value
     ///
-    /// This function is used for implementing `GraphQLType` for `i32` for all
+    /// This function is used for implementing `GraphQLValue` for `i32` for all
     /// scalar values. Implementations should convert all supported integer
     /// types with 32 bit or less to an integer if requested.
     fn as_int(&self) -> Option<i32>;
 
     /// Convert the given scalar value into a string value
     ///
-    /// This function is used for implementing `GraphQLType` for `String` for all
+    /// This function is used for implementing `GraphQLValue` for `String` for all
     /// scalar values
     fn as_string(&self) -> Option<String>;
 
     /// Convert the given scalar value into a string value
     ///
-    /// This function is used for implementing `GraphQLType` for `String` for all
+    /// This function is used for implementing `GraphQLValue` for `String` for all
     /// scalar values
     fn as_str(&self) -> Option<&str>;
 
     /// Convert the given scalar value into a float value
     ///
-    /// This function is used for implementing `GraphQLType` for `f64` for all
+    /// This function is used for implementing `GraphQLValue` for `f64` for all
     /// scalar values. Implementations should convert all supported integer
     /// types with 64 bit or less and all floating point values with 64 bit or
     /// less to a float if requested.
@@ -217,7 +229,7 @@ pub trait ScalarValue:
 
     /// Convert the given scalar value into a boolean value
     ///
-    /// This function is used for implementing `GraphQLType` for `bool` for all
+    /// This function is used for implementing `GraphQLValue` for `bool` for all
     /// scalar values.
     fn as_boolean(&self) -> Option<bool>;
 }